refactor: arrange file sturcture
