+++ xmind/fileshrink/__init__.py
