+++ xmind/fileshrink/pngquant/COPYRIGHT
