+++ xmind/fileshrink/pngquant/Drag PNG here to reduce palette automatically.bat
