+++ xmind/fileshrink/pngquant/Drag PNG here to reduce palette to 256.bat
