+++ xmind/fileshrink/pngquant/README.txt
