+++ xmind/fileshrink/pngquant/pngquant.exe
